requirements.txt
zeroentropy==0.1.0a6
python-dotenv==1.1.1
streamlit==1.32.0
streamlit-file-browser==0.1.0
pandas==2.1.4
Pillow==10.1.0
openai>=1.0.0
